Correct circular dependency
